fix: add api for sbbtc-apr
